Daemon/#156 Added `inertia [remote] token` cmd - Removed `inertia [remote] user token` - This only includes the admin privilege token
