+++ cmd/deploy.go
@@ -85,7 +85,6 @@ Run 'inertia [remote] init' to gather this information.`,
 		user.AddCommand(deepCopy(cmdDeploymentRemoveUser))
 		user.AddCommand(deepCopy(cmdDeploymentResetUsers))
 		user.AddCommand(deepCopy(cmdDeploymentListUsers))
-		user.AddCommand(deepCopy(cmdDeploymentGenerateToken))
 		cmd.AddCommand(user)
 
 		cmd.AddCommand(deepCopy(cmdDeploymentSSH))
@@ -105,6 +104,7 @@ Run 'inertia [remote] init' to gather this information.`,
 
 		cmd.AddCommand(deepCopy(cmdDeploymentInit))
 		cmd.AddCommand(deepCopy(cmdDeploymentReset))
+		cmd.AddCommand(deepCopy(cmdDeploymentToken))
 
 		remove := deepCopy(cmdDeploymentRemove)
 		cmd.AddCommand(remove)
@@ -528,5 +528,39 @@ directory (~/inertia) from your remote host.`,
 			log.Fatal(err)
 		}
 		println("Inertia and related daemon removed.")
+	},
+}
+
+var cmdDeploymentToken = &cobra.Command{
+	Use:   "token",
+	Short: "Generate tokens associated with permission levels for admin to share.",
+	Long:  `Generate tokens associated with permission levels for team leads to share`,
+	Run: func(cmd *cobra.Command, args []string) {
+		remoteName := strings.Split(cmd.Parent().Use, " ")[0]
+		deployment, _, err := local.GetClient(remoteName, configFilePath, cmd)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		resp, err := deployment.Token()
+		if err != nil {
+			log.Fatal(err)
+		}
+		defer resp.Body.Close()
+
+		body, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		switch resp.StatusCode {
+		case http.StatusOK:
+			fmt.Printf("New token: %s\n", string(body))
+		case http.StatusForbidden:
+			fmt.Printf("(Status code %d) Bad auth:\n%s\n", resp.StatusCode, string(body))
+		default:
+			fmt.Printf("(Status code %d) Unknown response from daemon:\n%s\n",
+				resp.StatusCode, body)
+		}
 	},
 }
