+++ cmd/users.go
@@ -176,37 +176,3 @@ var cmdDeploymentListUsers = &cobra.Command{
 		}
 	},
 }
-
-var cmdDeploymentGenerateToken = &cobra.Command{
-	Use:   "token",
-	Short: "Generate tokens associated with permission levels for admin to share.",
-	Long:  `Generate tokens associated with permission levels for team leads to share`,
-	Run: func(cmd *cobra.Command, args []string) {
-		remoteName := strings.Split(cmd.Parent().Parent().Use, " ")[0]
-		client, _, err := local.GetClient(remoteName, configFilePath, cmd)
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		resp, err := client.GenerateToken()
-		if err != nil {
-			log.Fatal(err)
-		}
-		defer resp.Body.Close()
-
-		body, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			log.WithError(err)
-		}
-
-		switch resp.StatusCode {
-		case http.StatusOK:
-			fmt.Printf("(Status code %d) %s\n", resp.StatusCode, string(body))
-		case http.StatusForbidden:
-			fmt.Printf("(Status code %d) Bad auth:\n%s\n", resp.StatusCode, string(body))
-		default:
-			fmt.Printf("(Status code %d) Unknown response from daemon:\n%s\n",
-				resp.StatusCode, body)
-		}
-	},
-}
